KC-74 Frontend styling (#42)
* Change styling from css to scss

* Add styling for dashboard

* Add styling for login

* Add styling for header

* Add styling for dialog

* Add styling for settings

* Alter model

* Remodell table in dashboard

* Style settings as card

* Use different font, style login as card

+++ dashboard/src/app/dashboard/dashboard.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {MatDialog, MatDialogConfig} from '@angular/material/dialog';
+import {MatDialog, MatDialogConfig, MatDialogRef} from '@angular/material/dialog';
 import {DialogComponent} from '../dialog/dialog.component';
 import {MatSnackBar} from '@angular/material/snack-bar';
 import {UsernamePasswordUrl} from '../models/username-password-url';
@@ -14,8 +14,8 @@ import {CrudService} from '../services/crud.service';
 })
 export class DashboardComponent implements OnInit {
 
-  header = ['id', 'username', 'password', 'url', 'Delete'];
-  entries: UsernamePasswordUrl[] = [];
+  displayedColumns: string[] = ['No.', 'Username', 'Password', 'Url', 'Delete'];
+  dataSource: UsernamePasswordUrl[] = [];
 
   constructor(
     private dialog: MatDialog,
@@ -28,8 +28,8 @@ export class DashboardComponent implements OnInit {
         if (resp.status === 200) {
           const body = JSON.parse(resp.body);
           body.forEach(item => {
-            const newEntry = new UsernamePasswordUrl(item.Id, item.Username, item.Password, item.Url);
-            this.entries.push(newEntry);
+            const newEntry = new UsernamePasswordUrl(item.Id, item.Username, item.Password, item.Url, '');
+            this.dataSource.push(newEntry);
           });
         } else {
           this.popOver.open(`${resp.status}`, '', {duration: 2000});
@@ -39,7 +39,7 @@ export class DashboardComponent implements OnInit {
           '', {duration: 5000});
       }
     );
-    console.log(this.entries);
+    console.log(this.dataSource);
   }
 
   ngOnInit() { }
@@ -55,22 +55,32 @@ export class DashboardComponent implements OnInit {
   }
 
   removeEntry(item) {
-    const index = this.entries.indexOf(item);
-    const body = new UsernameUrl(item.username, item.url);
-    this.crudService.deletePassword(body).subscribe(
-      resp => {
-        this.entries.splice(index, 1);
-        this.popOver.open('Deleted!', '', {duration: 2000});
-      }, error => {
-        this.popOver.open(`Something went wrong! If this error persists, please contact us with the following error: ${error.error}`,
-          '', {duration: 5000});
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe(
+      bool => {
+        if (bool) {
+          const index = this.dataSource.indexOf(item);
+          const body = new UsernameUrl(item.username, item.url);
+          this.crudService.deletePassword(body).subscribe(
+            resp => {
+              this.dataSource.splice(index, 1);
+              this.popOver.open('Deleted!', '', {duration: 2000});
+            }, error => {
+              this.popOver.open(`Something went wrong! If this error persists, please contact us with the following error: ${error.error}`,
+                '', {duration: 5000});
+            }
+          );
+        }
       }
     );
   }
 
   copyToClipboard(item) {
-    const index = this.entries.indexOf(item);
-    const password = this.entries[index].password;
+    const index = this.dataSource.indexOf(item);
+    const password = this.dataSource[index].password;
     const selBox = document.createElement('textarea');
     selBox.style.position = 'fixed';
     selBox.style.left = '0';
@@ -95,7 +105,7 @@ export class DashboardComponent implements OnInit {
     this.crudService.addPassword(newEntry).subscribe(
       resp => {
         if (resp.status === 200) {
-          this.entries.push(newEntry);
+          this.dataSource.push(newEntry);
           this.popOver.open('Saved', '', {duration: 2000});
         } else {
           this.popOver.open(`${resp.status}`, '', {duration: 2000});
@@ -105,6 +115,21 @@ export class DashboardComponent implements OnInit {
           '', {duration: 5000});
       }
     );
+  }
+
+}
+
+@Component({
+  selector: 'app-confirmation-dialog',
+  templateUrl: './confirmation.component.html',
+})
+export class ConfirmationDialogComponent {
+
+  constructor(
+    public dialogRef: MatDialogRef<ConfirmationDialogComponent>) {}
+
+  onNoClick(): void {
+    this.dialogRef.close();
   }
 
 }
